--- mage_ai/api/resources/mixins/TokenEncodable.py
@@ -1,11 +1,0 @@
-import jwt
-
-JWT_SECRET = 'materia'
-
-
-class TokenEncodable():
-    def encode_token(self, token, expires):
-        return jwt.encode({
-            'expires': expires.timestamp(),
-            'token': token,
-        }, JWT_SECRET, algorithm='HS256')
