+++ mage_ai/authentication/oauth2.py
@@ -1,6 +1,11 @@
 from datetime import datetime, timedelta
 from mage_ai.orchestration.db.models import Oauth2AccessToken, Oauth2Application, User
+from typing import Dict
+import jwt
 import secrets
+
+JWT_ALGORITHM = 'HS256'
+JWT_SECRET = 'materia'
 
 
 def generate_access_token(user: User, application: Oauth2Application = None) -> Oauth2AccessToken:
@@ -25,3 +30,14 @@ def generate_access_token(user: User, application: Oauth2Application = None) -> 
 
 def get_access_token(token: str) -> Oauth2AccessToken:
     return Oauth2AccessToken.query.filter(Oauth2AccessToken.token == token).first()
+
+
+def encode_token(token: str, expires: datetime) -> str:
+    return jwt.encode({
+        'expires': expires.timestamp(),
+        'token': token,
+    }, JWT_SECRET, algorithm=JWT_ALGORITHM)
+
+
+def decode_token(encoded_token: str) -> Dict:
+    return jwt.decode(encoded_token, key=JWT_SECRET, algorithms=[JWT_ALGORITHM], verify=True)
