+++ mage_ai/orchestration/db/errors.py
@@ -1,5 +1,14 @@
 from mage_ai.errors.base import MageBaseException
+from typing import Dict
 
 
 class ValidationError(MageBaseException):
-    pass
+    def __init__(self, error: str, metadata: Dict):
+        self.error = error
+        self.metadata = metadata
+
+    def to_dict(self):
+        return dict(
+            error=self.error,
+            metadata=self.metadata,
+        )
