+++ mage_ai/orchestration/db/models.py
@@ -17,6 +17,7 @@ from sqlalchemy.orm.collections import InstrumentedList
 from sqlalchemy.sql import func
 from typing import Dict, List
 import enum
+import json
 import re
 import uuid
 
@@ -136,12 +137,15 @@ class User(BaseModel):
     oauth2_access_tokens = relationship('Oauth2AccessToken', back_populates='user')
 
     @validates('email')
-    def validate_email(self, key, email):
-        if email:
+    def validate_email(self, key, value):
+        if value:
             regex = re.compile(r"([-!#-'*+/-9=?A-Z^-~]+(\.[-!#-'*+/-9=?A-Z^-~]+)*|\"([]!#-[^-~ \t]|(\\[\t -~]))+\")@([-!#-'*+/-9=?A-Z^-~]+(\.[-!#-'*+/-9=?A-Z^-~]+)*|\[[\t -Z^-~]*])")
-            if not re.fullmatch(regex, email):
-                raise ValidationError('Email address is invalid.')
-        return email
+            if not re.fullmatch(regex, value):
+                raise ValidationError('Email address is invalid.', metadata=dict(
+                    key=key,
+                    value=value,
+                ))
+        return value
 
 
 class Oauth2Application(BaseModel):
