+++ mage_ai/server/api/constants.py
@@ -2,3 +2,4 @@ ENDPOINTS_BYPASS_OAUTH_CHECK = []
 HEADER_API_KEY = 'X-API-KEY'
 HEADER_OAUTH_TOKEN = 'OAUTH-TOKEN'
 OAUTH_CLIENT_ID_KEY = 'OAUTH_CLIENT_ID'
+URL_PARAMETER_API_KEY = 'api_key'
